Make native auth handoff route testable
